Revert "Implement `snapcraft push` (#653)"
This reverts commit b174930109a79220d6a9180f75f195b2509b2033.
