+++ snapcraft/main.py
@@ -33,7 +33,6 @@ Usage:
   snapcraft [options] logout
   snapcraft [options] register <snap-name>
   snapcraft [options] upload <snap-file>
-  snapcraft [options] push <snap-file> [--release <channels>]
   snapcraft [options] release <snap-name> <revision> <channel>
   snapcraft [options] list-plugins
   snapcraft [options] tour [<directory>]
@@ -71,9 +70,6 @@ Options specific to snapping:
   -o <snap-file>, --output <snap-file>  used in case you want to rename the
                                         snap.
 
-Options specific to store interaction:
-  --release <channels>  Comma separated list of channels to release to.
-
 The available commands are:
   help         Obtain help for a certain plugin or topic
   init         Initialize a snapcraft project.
@@ -83,9 +79,7 @@ The available commands are:
   register     Register the package name in the store.
   tour         Setup the snapcraft examples tour in the specified directory,
                or ./snapcraft-tour/.
-  push         Pushes and optionally releases a snap to the Ubuntu Store.
-  upload       DEPRECATED Upload a snap to the Ubuntu Store. The push command
-               supersedes this command.
+  upload       Upload a snap to the Ubuntu Store.
   release      Release a revision of a snap to a specific channel.
 
 The available lifecycle commands are:
@@ -278,7 +272,7 @@ def _run_clean(args, project_options):
 
 
 def _is_store_command(args):
-    commands = ('register', 'upload', 'release', 'push')
+    commands = ('register', 'upload', 'release')
     return any(args.get(command) for command in commands)
 
 
@@ -286,17 +280,10 @@ def _run_store_command(args):
     if args['register']:
         snapcraft.register(args['<snap-name>'])
     elif args['upload']:
-        logger.warning('DEPRECATED: Use `push` instead of `upload`')
-        snapcraft.push(args['<snap-file>'])
-    elif args['push']:
-        if args['--release']:
-            release_channels = args['--release'].split(',')
-        else:
-            release_channels = []
-        snapcraft.push(args['<snap-file>'], release_channels)
+        snapcraft.upload(args['<snap-file>'])
     elif args['release']:
         snapcraft.release(
-            args['<snap-name>'], args['<revision>'], [args['<channel>']])
+            args['<snap-name>'], args['<revision>'], args['<channel>'])
 
 
 if __name__ == '__main__':  # pragma: no cover
